Django/StorageSystemAPI/src/requirements.txt
Babel==2.8.0
Django==2.2.2
django-phonenumber-field==4.0.0
phonenumbers==8.12.6
pytz==2020.1
sqlparse==0.3.1
